requirements.txt
SalesGPT==0.1.1
Flask==2.3.3
Flask-Cors==4.0.0
langchain== 0.1.0
google-generativeai==0.3.2
python-dotenv==1.0.1

